+++ wildRumble/src/de/silentinfiltration/game/systems/RenderSystem.java
@@ -57,7 +57,7 @@ public class RenderSystem extends BaseSystem {
 	}
 
 	@Override
-	public void update(long dt, int entity) throws ComponentNotFoundEx {
+	public void update(double dt, int entity) throws ComponentNotFoundEx {
 		
 		
 	}
